+++ dangerfile.flow.js
@@ -1,0 +1,8 @@
+// This file isn't used anywhere, but it is used as a part of `yarn flow check`
+// to validate that flow typings work correctly for JS Dangerfiles
+
+// @flow
+
+import { danger } from "danger"
+
+danger.github.pr