+++ source/ci_source/ci_source_helpers.ts
@@ -1,4 +1,4 @@
-import { Env, RepoMetaData } from "./ci_source"
+import { Env } from "./ci_source"
 import { GitHubAPI } from "../platforms/github/GitHubAPI"
 import { GitHubPRDSL } from "../dsl/GitHubDSL"
 import * as find from "lodash.find"
@@ -6,6 +6,7 @@ import {
   BitBucketServerAPI,
   bitbucketServerRepoCredentialsFromEnv,
 } from "../platforms/bitbucket_server/BitBucketServerAPI"
+import { RepoMetaData } from "../dsl/BitBucketServerDSL"
 
 /**
  * Validates that all ENV keys exist and have a length