+++ source/dsl/DangerDSL.ts
@@ -1,8 +1,10 @@
+// Please don't have includes in here that aren't inside the DSL folder, or the d.ts/flow defs break
+
 import { GitDSL, GitJSONDSL } from "../dsl/GitDSL"
 import { GitHubDSL } from "../dsl/GitHubDSL"
 import { BitBucketServerDSL, BitBucketServerJSONDSL } from "../dsl/BitBucketServerDSL"
 import { DangerUtilsDSL } from "./DangerUtilsDSL"
-import { CliArgs } from "../runner/cli-args"
+import { CliArgs } from "../dsl/cli-args"
 
 /**
  * The shape of the JSON passed between Danger and a subprocess. It's built
@@ -96,17 +98,30 @@ export interface DangerDSLType {
    *  GitHub user identities and some useful utility functions
    *  for displaying links to files.
    *
+   * Strictly speaking, `github` is a nullable type, if you are not using
+   * GitHub then it will be undefined. For the DSL convience sake though, it
+   * is classed as non-nullable
+   *
    *  Provides an authenticated API so you can work directly
    *  with the GitHub API. This is an instance of the "@ocktokit/rest" npm
    *  module.
    *
    *  Finally, if running through Peril on an event other than a PR
-   *  this is the full JSON from the webhook. You can find the full
-   *  typings for those webhooks [at github-webhook-event-types](https://github.com/orta/github-webhook-event-types).
+   *  this is the full JSON from the webhook. [github-webhook-event-types](https://github.com/orta/github-webhook-event-types) has the full
+   *  typings for those webhooks.
    */
-  readonly github?: GitHubDSL
+  readonly github: GitHubDSL
 
-  readonly bitbucket_server?: BitBucketServerDSL
+  /**
+   *  The BitBucket Server metadata. This covers things like PR info,
+   *  comments and reviews on the PR, related issues, commits, comments
+   *  and activities.
+   *
+   *  Strictly speaking, `bitbucket_server` is a nullable type, if you are using
+   *  GitHub then it will be undefined. For the DSL convience sake though, it
+   *  is classed as non-nullable
+   */
+  readonly bitbucket_server: BitBucketServerDSL
 
   /**
    * Functions which are globally useful in most Dangerfiles. Right