+++ CHANGELOG.md
@@ -13,8 +13,62 @@
 
 ## Master
 
+* []
+
+## 3.2.0
+
+=======
+
+* Add BitBucket Server support.
+
+  To use Danger JS with BitBucket Server: you'll need to create a new account for Danger to use,
+  then set the following environment variables on your CI:
+
+  * `DANGER_BITBUCKETSERVER_HOST` = The root URL for your server, e.g. `https://bitbucket.mycompany.com`.
+  * `DANGER_BITBUCKETSERVER_USERNAME` = The username for the account used to comment.
+  * `DANGER_BITBUCKETSERVER_PASSWORD` = The password for the account used to comment.
+
+  Then you will have a fully fleshed out `danger.bitbucket_server` object in your Dangerfile to work with,
+  for example:
+
+  ```ts
+  import { danger, warn } from "danger"
+
+  if (danger.bitbucket_server.pr.title.includes("WIP")) {
+    warn("PR is considered WIP")
+  }
+  ```
+
+  The DSL is fully fleshed out, you can see all the details inside the [Danger JS Reference][ref],
+  but the summary is:
+
+  ```ts
+  danger.bitbucket_server.
+    /** The pull request and repository metadata */
+    metadata: RepoMetaData
+    /** The related JIRA issues */
+    issues: JIRAIssue[]
+    /** The PR metadata */
+    pr: BitBucketServerPRDSL
+    /** The commits associated with the pull request */
+    commits: BitBucketServerCommit[]
+    /** The comments on the pull request */
+    comments: BitBucketServerPRActivity[]
+    /** The activities such as OPENING, CLOSING, MERGING or UPDATING a pull request */
+    activities: BitBucketServerPRActivity[]
+  ```
+
+  You can see more in the docs for [Danger + BitBucket Server](http://danger.systems/js/usage/bitbucket_server.html).
+
+  * [@azz][]
+
+* Don't check for same user ID on comment when running as a GitHub App. [@tibdex][]
+
+## 3.1.8
+
+* Improvements to the Flow definition file. [@orta][]
+* Improve path generator for danger-runner.
 * Update the PR DSL to include bots. [@orta][]
-* Add BitBucket server support. [@azz][]
 
 ## 3.1.7
 
@@ -908,3 +962,5 @@ Not usable for others, only stubs of classes etc. - [@orta][]
 [@hongrich]: https://github.com/hongrich
 [@peterjgrainger]: https://github.com/peterjgrainger
 [@azz]: https://github.com/azz
+[@mifi]: https://github.com/ionutmiftode
+[ref]: http://danger.systems/js/reference.html